chore: split library functions (WIP)
