--- src/components/utils.ts
@@ -1,1 +1,0 @@
-export const nothing = {}
